osmo-bts-trx: implement CSD scheduling support
* enlarge the maximum burst buffer size to 24 * (2 * 58) bytes;
* enlarge per-l1cs Uplink burst mask to hold up to 32 bits;
* enlarge per-l1cs Uplink meas ring buffer to 24 entries;
* add new meas modes: SCHED_MEAS_AVG_M_{S22N22,S24N22};

Depends: libosmocore.git Ib482817b5f6a4e3c7299f6e0b3841143b60fc93d
Related: OS#1572

+++ src/osmo-bts-trx/sched_lchan_pdtch.c
@@ -44,7 +44,7 @@ int rx_pdtch_fn(struct l1sched_ts *l1ts, const struct trx_ul_burst_ind *bi)
 	struct l1sched_chan_state *chan_state = &l1ts->chan_state[bi->chan];
 	sbit_t *burst, *bursts_p = chan_state->ul_bursts;
 	uint32_t first_fn;
-	uint8_t *mask = &chan_state->ul_mask;
+	uint32_t *mask = &chan_state->ul_mask;
 	struct l1sched_meas_set meas_avg;
 	uint8_t l2[EGPRS_0503_MAX_BYTES];
 	int n_errors = 0;
